move all channel codec tests to `ChannelCodecsSpec`
