+++ eclair-core/src/test/scala/fr/acinq/eclair/db/SqliteChannelsDbSpec.scala
@@ -21,6 +21,7 @@ import fr.acinq.eclair.TestConstants
 import fr.acinq.eclair.db.sqlite.SqliteUtils.{getVersion, using}
 import fr.acinq.eclair.db.sqlite.{SqliteChannelsDb, SqlitePendingRelayDb}
 import fr.acinq.eclair.wire.ChannelCodecs.stateDataCodec
+import fr.acinq.eclair.wire.ChannelCodecsSpec
 import org.scalatest.FunSuite
 import org.sqlite.SQLiteException
 import scodec.bits.ByteVector
@@ -39,7 +40,7 @@ class SqliteChannelsDbSpec extends FunSuite {
     val db = new SqliteChannelsDb(sqlite)
     new SqlitePendingRelayDb(sqlite) // needed by db.removeChannel
 
-    val channel = ChannelStateSpec.normal
+    val channel = ChannelCodecsSpec.normal
 
     val commitNumber = 42
     val paymentHash1 = ByteVector32.Zeroes
@@ -78,7 +79,7 @@ class SqliteChannelsDbSpec extends FunSuite {
     }
 
     // insert 1 row
-    val channel = ChannelStateSpec.normal
+    val channel = ChannelCodecsSpec.normal
     val data = stateDataCodec.encode(channel).require.toByteArray
     using(sqlite.prepareStatement("INSERT INTO local_channels VALUES (?, ?)")) { statement =>
       statement.setBytes(1, channel.channelId.toArray)
